requirements.txt
requests==2.31.0
polars==0.20.8
apache-airflow[postgres,celery,amazon]==2.8.1
psycopg2-binary==2.9.1
Flask-Session==0.5.0